v154

+++ libft/ft_atoi.c
@@ -12,8 +12,8 @@
 
 int	ft_atoi(const char *str)
 {
-	int	res;
-	int	sign;
+	long long int	res;
+	int				sign;
 
 	sign = 1;
 	res = 0;
@@ -26,5 +26,5 @@ int	ft_atoi(const char *str)
 	}
 	while (*str >= '0' && *str <= '9')
 		res = res * 10 + *str++ - '0';
-	return (res * sign);
+	return ((int)res * sign);
 }
